backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
dj-database-url==3.1.2
Django==5.2.11
django-cors-headers==4.9.0
django-filter==25.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.15.0
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.1.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
shapely>=2.0
whitenoise==6.12.0
celery[redis]>=5.4
redis>=5.0
